tso: move GetLeader to util package
