+++ util/util_test.go
@@ -11,15 +11,24 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package server
+package util
 
 import (
+	"encoding/json"
+	"flag"
+	"testing"
 	"time"
 
 	"github.com/ngaut/go-zookeeper/zk"
 	"github.com/ngaut/zkhelper"
 	. "github.com/pingcap/check"
 )
+
+var testZKAddr = flag.String("zk", "127.0.0.1:2181", "test zookeeper address")
+
+func TestUtil(t *testing.T) {
+	TestingT(t)
+}
 
 var _ = Suite(&testUtilSuite{})
 
@@ -36,8 +45,6 @@ func (s *testUtilSuite) SetUpSuite(c *C) {
 
 	s.rootPath = "/zk/tso_util_test"
 
-	_, err = zkhelper.CreateRecursive(s.zkConn, s.rootPath, "", 0, zk.WorldACL(zkhelper.PERM_DIRECTORY))
-	c.Assert(err, IsNil)
 }
 
 func (s *testUtilSuite) TearDownSuite(c *C) {
@@ -49,31 +56,45 @@ func (s *testUtilSuite) TearDownSuite(c *C) {
 	}
 }
 
-func (s *testUtilSuite) TestTimestamp(c *C) {
+func (s *testUtilSuite) TestLeader(c *C) {
 	conn, err := zkhelper.ConnectToZkWithTimeout(*testZKAddr, time.Second)
 	c.Assert(err, IsNil)
 	defer conn.Close()
 
-	tbl := []int64{
-		1, 100, 1000,
-	}
+	leaderPath := getLeaderPath(s.rootPath)
+	conn.Delete(leaderPath, -1)
 
-	for _, t := range tbl {
-		err = saveTimestamp(conn, s.rootPath, t)
-		c.Assert(err, IsNil)
-
-		n, err := loadTimestamp(conn, s.rootPath)
-		c.Assert(err, IsNil)
-		c.Assert(n, Equals, t)
-	}
-
-	// test error
-	_, err = loadTimestamp(conn, "error_root_path")
+	_, err = GetLeader(conn, s.rootPath)
 	c.Assert(err, NotNil)
 
-	_, err = conn.Set(getTimestampPath(s.rootPath), []byte{}, -1)
+	_, _, err = GetWatchLeader(conn, s.rootPath)
+	c.Assert(err, NotNil)
+
+	_, err = zkhelper.CreateRecursive(conn, leaderPath, "", 0, zk.WorldACL(zkhelper.PERM_FILE))
 	c.Assert(err, IsNil)
 
-	_, err = loadTimestamp(conn, s.rootPath)
+	_, err = GetLeader(conn, s.rootPath)
 	c.Assert(err, NotNil)
+
+	_, _, err = GetWatchLeader(conn, s.rootPath)
+	c.Assert(err, NotNil)
+
+	addr := "127.0.0.1:1234"
+	m := map[string]interface{}{
+		"Addr": addr,
+	}
+
+	data, err := json.Marshal(m)
+	c.Assert(err, IsNil)
+
+	_, err = conn.Set(leaderPath, data, -1)
+	c.Assert(err, IsNil)
+
+	v, err := GetLeader(conn, s.rootPath)
+	c.Assert(err, IsNil)
+	c.Assert(v, Equals, addr)
+
+	v, _, err = GetWatchLeader(conn, s.rootPath)
+	c.Assert(err, IsNil)
+	c.Assert(v, Equals, addr)
 }
